+++ src/config/apply.rs
@@ -89,6 +89,18 @@ pub fn apply_config(
         args.no_charts = no_charts;
     }
 
+    if !is_cli(matches, "no_ua")
+        && let Some(no_ua) = config.no_ua
+    {
+        args.no_ua = no_ua;
+    }
+
+    if !is_cli(matches, "authorized")
+        && let Some(authorized) = config.authorized
+    {
+        args.authorized = authorized;
+    }
+
     if !is_cli(matches, "tmp_path")
         && let Some(path) = config.tmp_path.clone()
     {