+++ src/http/tests.rs
@@ -18,6 +18,8 @@ fn base_args(url: String) -> Result<TesterArgs, String> {
         method: HttpMethod::Get,
         url: Some(url),
         headers: vec![],
+        no_ua: false,
+        authorized: false,
         data: String::new(),
         target_duration: positive_u64(1)?,
         expected_status_code: 200,